updated groups schema to include permissions as well as creating a permissions helper function in auth package
