+++ Makefile
@@ -5,3 +5,6 @@ dev:
 
 dev-build:
 	GO_MOD=go.mod docker compose up --build
+
+pg:
+	docker volume prune; GO_MOD=go.mod docker compose up --build
